√ vagrant: improve plugin setup idempotency
Introduce globbing in creates= and removes= conditions, with a custom
version of the command module.

Note that this feature should be part of Ansible 1.9.3+, and even
extended in Ansible 2.x.

This raise the question #26 (package a specific version of Ansible) if
we want to get rid of "disp_" modules over time...